+++ src/OpenTelemetry.Instrumentation.Wcf/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 * Update OpenTelemetry SDK version to `1.6.0`.
   ([#1344](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1344))
+* Fixed span hierarchy when hosted in ASP.NET
+  ([#1342](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1342))
+* **Breaking Change** `TelemetryClientMessageInspector` and `TelemetryDispatchMessageInspector`
+  changed from public to internal
+  ([#1376](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1376))
 
 ## 1.0.0-rc.12